PC-41 rename project and app
