+++ srcs/backend/transcendence/.gitignore
@@ -1,0 +1,3 @@
+transcendence/__pycache__/
+app/__pycache__
+env/
