+++ srcs/backend/transcendence/entrypoint.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 python manage.py makemigrations
-python manage.py makemigrations user
+python manage.py makemigrations app
 python manage.py migrate
 
 exec "$@"
