+++ srcs/backend/transcendence/templates/user/delete_account.html
