+++ srcs/backend/transcendence/templates/user/logout.html
