+++ srcs/backend/transcendence/templates/user/manage_account.html
