+++ srcs/backend/transcendence/templates/user/register.html
