+++ srcs/backend/transcendence/transcendence/cors_middleware.py
