+++ srcs/backend/transcendence/transcendence/routing.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from app.consumers import UserConsumer
+
+websocket_urlpatterns = [
+    path("ws/app/", UserConsumer.as_asgi()),
+]
