+++ srcs/backend/transcendence/transcendence/urls.py
@@ -18,5 +18,5 @@ from django.contrib import admin
 from django.urls import path, include
 
 urlpatterns = [
-    path('user/', include('user.urls')),
+    path('app/', include('app.urls')),
 ]
