+++ srcs/backend/transcendence/transcendence/wsgi.py
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'user_service.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'transcendence.settings')
 
 application = get_wsgi_application()
