--- srcs/backend/user_service/.gitignore
@@ -1,3 +1,0 @@
-user_service/__pycache__/
-user/__pycache__
-env/
