--- srcs/backend/user_service/user_service/routing.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from user.consumers import UserConsumer
-
-websocket_urlpatterns = [
-    path("ws/user/", UserConsumer.as_asgi()),
-]
