+++ srcs/frontend/conf/nginx.conf
@@ -39,22 +39,15 @@ http {
             try_files $uri $uri/ /index.html;
         }
 
-        location /ws/pong/ {
-            proxy_pass http://pong_service:8000/ws/pong/;
+        location /ws/app/ {
+            proxy_pass http://transcendence:8000/ws/app/;
             proxy_http_version 1.1;
             proxy_set_header Upgrade $http_upgrade;
             proxy_set_header Connection "Upgrade";
         }
 
-        location /ws/user/ {
-            proxy_pass http://user-service:8000/ws/user/;
-            proxy_http_version 1.1;
-            proxy_set_header Upgrade $http_upgrade;
-            proxy_set_header Connection "Upgrade";
-        }
-
-        location /user/ {
-            proxy_pass http://user-service:8000/user/;
+        location /app/ {
+            proxy_pass http://transcendence:8000/app/;
         }
     }
 }
